fix: address graceful shutdown review feedback
