+++ pkg/wire/http_runtime_binding.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"errors"
 	"net/http"
+	"time"
 
 	"github.com/portpowered/infinite-you/pkg/services/costs"
+	costscli "github.com/portpowered/infinite-you/pkg/services/costs/transports/cli"
 	costshttp "github.com/portpowered/infinite-you/pkg/services/costs/transports/http"
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factorydefinitionshttp "github.com/portpowered/infinite-you/pkg/services/factory_definitions/transports/http"
@@ -14,6 +16,7 @@ import (
 	factorysessionshttp "github.com/portpowered/infinite-you/pkg/services/factory_sessions/transports/http"
 	factorysessionwire "github.com/portpowered/infinite-you/pkg/services/factory_sessions/wire"
 	factoryvisualization "github.com/portpowered/infinite-you/pkg/services/factory_visualization"
+	visualizationcli "github.com/portpowered/infinite-you/pkg/services/factory_visualization/transports/cli"
 	factoryvisualizationhttp "github.com/portpowered/infinite-you/pkg/services/factory_visualization/transports/http"
 	modelshttp "github.com/portpowered/infinite-you/pkg/services/models/transports/http"
 	providersessionshttp "github.com/portpowered/infinite-you/pkg/services/provider_sessions/transports/http"
@@ -22,9 +25,30 @@ import (
 	workhttp "github.com/portpowered/infinite-you/pkg/services/work/transports/http"
 	workersessionshttp "github.com/portpowered/infinite-you/pkg/services/worker_sessions/transports/http"
 	transporthttp "github.com/portpowered/infinite-you/pkg/transports/http"
+	generatedhttpclient "github.com/portpowered/infinite-you/pkg/transports/http/client"
 	apisurface "github.com/portpowered/infinite-you/pkg/transports/mapping"
 	factorysessionmapping "github.com/portpowered/infinite-you/pkg/transports/mapping/factorysession"
 )
+
+const metricsCLIHTTPTimeout = 5 * time.Minute
+
+func provideCostsCLI() costscli.Operation {
+	return costscli.NewOperation(func(server string) (costscli.Client, error) {
+		return generatedhttpclient.NewClientWithResponses(
+			server,
+			generatedhttpclient.WithHTTPClient(&http.Client{Timeout: costscli.DefaultRequestTimeout}),
+		)
+	})
+}
+
+func provideMetricsCLI() visualizationcli.Operation {
+	return visualizationcli.NewOperation(func(server string) (visualizationcli.Client, error) {
+		return generatedhttpclient.NewClientWithResponses(
+			server,
+			generatedhttpclient.WithHTTPClient(&http.Client{Timeout: metricsCLIHTTPTimeout}),
+		)
+	})
+}
 
 // httpRuntimeBinding is the Wire-owned operation that builds the owner HTTP
 // adapters for one opened runtime and returns the generated route shell.
@@ -242,7 +266,7 @@ func newHTTPRuntimeServer(
 	if opened.Cancellation != nil {
 		shutdown = opened.Cancellation.Cancel
 	}
-	return transporthttp.NewServerWithRecordingsAndMetricsAndCostsAndShutdown(
+	return transporthttp.NewServerWithRecordingsAndMetricsAndCosts(
 		recordingsAdapter, sessionsHandler, workHandler, modelsHandler, providerSessionsHTTP,
 		factoryDefinitionsHandler, opened.Logger, metricsHandler, costsHandler, shutdown, workerSessionsHandler,
 	).Handler()
