+++ tests/functional/transport/http/server/routing_test.go
@@ -8,6 +8,7 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/portpowered/infinite-you/internal/contractinventory"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
@@ -24,8 +25,13 @@ func TestAPIRoutesEveryOpenAPIOperationToNon404Handler(t *testing.T) {
 	defer ctx.server.Stop(t)
 
 	client := &http.Client{Timeout: routingReachabilityRequestTimeout}
+	var shutdownOperation *contractinventory.Operation
 	for _, operation := range inventory.Operations {
 		operation := operation
+		if operation.OperationID == "shutdownServer" {
+			shutdownOperation = &operation
+			continue
+		}
 		t.Run(operation.OperationID, func(t *testing.T) {
 			request, err := ctx.safeRequest(operation)
 			if err != nil {
@@ -67,6 +73,36 @@ func TestAPIRoutesEveryOpenAPIOperationToNon404Handler(t *testing.T) {
 	if len(inventory.Operations) == 0 {
 		t.Fatal("OpenAPI operation inventory is empty")
 	}
+	if shutdownOperation == nil {
+		t.Fatal("OpenAPI operation inventory does not contain shutdownServer")
+	}
+
+	// Shutdown cancels the shared functional server after acknowledging the
+	// request, so it must be the final operation exercised by this fixture.
+	t.Run(shutdownOperation.OperationID, func(t *testing.T) {
+		request, err := ctx.safeRequest(*shutdownOperation)
+		if err != nil {
+			t.Fatalf("build safe request for %s %s (%s): %v", shutdownOperation.Method, shutdownOperation.Path, shutdownOperation.OperationID, err)
+		}
+
+		response, err := client.Do(request)
+		if err != nil {
+			t.Fatalf("%s %s (%s): %v", shutdownOperation.Method, request.URL.String(), shutdownOperation.OperationID, err)
+		}
+		defer response.Body.Close()
+		if response.StatusCode != http.StatusAccepted {
+			body, _ := io.ReadAll(response.Body)
+			t.Fatalf("%s %s (%s) status = %d, want %d: %s", shutdownOperation.Method, request.URL.String(), shutdownOperation.OperationID, response.StatusCode, http.StatusAccepted, strings.TrimSpace(string(body)))
+		}
+
+		timer := time.NewTimer(routingReachabilityRequestTimeout)
+		defer timer.Stop()
+		select {
+		case <-ctx.server.Done():
+		case <-timer.C:
+			t.Fatal("shutdownServer acknowledged but functional server did not stop")
+		}
+	})
 }
 
 func assertModelCacheNotFoundResponse(t *testing.T, response *http.Response) {
